BAEL-6310 - update project location
