+++ jackson-modules/jackson-jr/pom.xml
@@ -9,9 +9,8 @@
 
     <parent>
         <groupId>com.baeldung</groupId>
-        <artifactId>parent-java</artifactId>
+        <artifactId>jackson-modules</artifactId>
         <version>0.0.1-SNAPSHOT</version>
-        <relativePath>../parent-java</relativePath>
     </parent>
 
     <dependencies>
@@ -34,5 +33,15 @@
         </dependency>
 
     </dependencies>
+
+    <build>
+        <finalName>jackson-jr</finalName>
+        <resources>
+            <resource>
+                <directory>src/main/resources</directory>
+                <filtering>true</filtering>
+            </resource>
+        </resources>
+    </build>
 
 </project>
