+++ jackson-modules/jackson-jr/src/main/java/com/baeldung/jacksonjr/Person.java
@@ -1,12 +1,12 @@
 package com.baeldung.jacksonjr;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
+import java.time.LocalDate;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.time.LocalDate;
 
 @Data
 @NoArgsConstructor
