+++ jackson-modules/pom.xml
@@ -21,6 +21,7 @@
         <module>jackson-core</module>
         <module>jackson-custom-conversions</module>
         <module>jackson-exceptions</module>
+        <module>jackson-jr</module>
     </modules>
 
     <dependencies>
